Added lots of new validation for adding/removing halls. For a certain event only the admin organizer can do it.

+++ src/main/java/fmi/eventmanager/Agendy/service/HallService.java
@@ -4,9 +4,14 @@ import fmi.eventmanager.Agendy.model.dto.Halls.CreateHallRequest;
 import fmi.eventmanager.Agendy.model.dto.Halls.HallResponse;
 import fmi.eventmanager.Agendy.model.entity.Event;
 import fmi.eventmanager.Agendy.model.entity.Hall;
+import fmi.eventmanager.Agendy.model.entity.Role;
+import fmi.eventmanager.Agendy.model.entity.User;
 import fmi.eventmanager.Agendy.repository.EventRepository;
 import fmi.eventmanager.Agendy.repository.HallRepository;
+import fmi.eventmanager.Agendy.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -15,15 +20,20 @@ public class HallService {
 
     private final HallRepository hallRepository;
     private final EventRepository eventRepository;
+    private final UserRepository userRepository;
 
-    public HallService(HallRepository hallRepository, EventRepository eventRepository) {
+    public HallService(HallRepository hallRepository, EventRepository eventRepository, UserRepository userRepository) {
         this.hallRepository = hallRepository;
         this.eventRepository = eventRepository;
+        this.userRepository = userRepository;
     }
 
-    public HallResponse createHall(Long eventId, CreateHallRequest request) {
+    public HallResponse createHall(Long userId, Long eventId, CreateHallRequest request) {
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new RuntimeException("Event not found"));
+        if (userId != event.getOrganizerId()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only organizers can add halls.");
+        }
         Hall hall = new Hall(event, request.getName(), request.getCapacity());
         return mapToResponse(hallRepository.save(hall));
     }
@@ -41,9 +51,12 @@ public class HallService {
         return responses;
     }
 
-    public void deleteHall(Long hallId) {
-        hallRepository.findById(hallId)
-                .orElseThrow(() -> new RuntimeException("Hall not found"));
+    public void deleteHall(Long userId, Long hallId) {
+        Hall hall = hallRepository.findById(hallId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Hall not found!"));
+        if (hall.getEvent().getOrganizerId() != userId) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only organizers of events can delete rooms");
+        }
         hallRepository.deleteById(hallId);
     }
